资产兑换页面功能实现； WMNC、MNC明细页面

+++ lib/src/main/java/com/alaer/lib/api/bean/TeamInfo.java
@@ -45,7 +45,7 @@ public class TeamInfo implements Serializable {
         public float minerActivityness;
         public float promotionActivityness;
         public float fireActivityness;
-        public double reputation;
+        public float reputation;
         public int contribution;
         public int level;
         public String createTime;
